[LION-27] Set up Local Database
This commit implements a Recipe Database, a Recipe Dao for searching that database,
a recipe repository to determine where to look (in this case, just to Recipe Database, though eventually also to Firebase)
and connects the RecipeModel to the Repository. Everything is updated to the Recipe Model through the LiveData class.